moved SYSTICK timer initialization step out from kernel.c

+++ src/api/kernel.h
@@ -14,7 +14,7 @@ extern "C" {
 
 extern const uint32_t zero_ref;
 
-extern uint8_t kernel_init();
+extern uint32_t kernel_init();
 extern void start_scheduler(void);
 extern uint8_t device_reset();
 extern uint8_t cpu_freq_update(uint32_t frequency);
